Add importModuleFromResources (#2)
Add an importModuleFromResources, which read the code from the resources
file, and didn't use the importlib so can be use with python < 3.1.
This doesn't enable import of module located in QtResources from the
python code, which will require a more complex implementation with hook
in python capi. This solve the #2 enhancement request.